Bug#29360763 WL#1075: FUNCTIONAL INDEX: INCORRECT ERROR MESSAGE REVEALS HIDDEN COLUMN NAME
When renaming a functional index to the same name as a different
functional index, the names of the hidden columns would end up
equal. The result was that the server returned an "Internal error"
message instead of the expected "Duplicate key name" since the
check for duplicate column names happens before the check for
duplicate index names. This patch changes the check for duplicate
hidden column names, so a more appropriate error can be reported.

Change-Id: I41147444b718d1a517211c5d29271f5487b9fd86