Bug#29368521: ASSERTION `!THD->IS_ERROR()' FAILED
If an error is raised when evaluating the ESCAPE clause of a LIKE
expression at resolve time, the error status is not propagated to the
caller. An assertion failed because the thd->is_error() flag was set
when resolution returned successfully.

The fix is to add a check for thd->is_error() after evaluating the
ESCAPE clause in Item_func_like::eval_escape_clause(), and return an
error code if an error has been raised.

Change-Id: I388ea1725fc33ce17ce6dd24b859a6cc62811197