WL #4168: Implement EXPLAIN ANALYZE [patch 2/2, actual implementation]
Implement EXPLAIN ANALYZE, by means of a new class template
TimingIterator<T> that wraps each iterator. NewIterator will create
a TimingIterator if needed, which times each Init() and Read() call.
EXPLAIN ANALYZE is then just a matter of running the query with the
result discarded, and running a regular EXPLAIN with the timing
information attached.

Change-Id: I07e3e537293212b4368f387670201a1568c8ea7a