Bug #27491547 PERFSCHEMA.INNODB_IS_DATA_LOCKS FAILS SPORADICALLY ON PB2
IActually the test no longer fails on Windows after the
Bug #28176910 PERFORMANCE_SCHEMA.DATA_LOCKS.OBJECT_INSTANCE_BEGIN CHANGES DEPENDING ON ORDER
got fixed.
However, the test and result file still needed to be updated because we've changed description
of the lock to contain ",REC_NOT_GAP" in case of locks which are on record only and not on the gap.

RB:21546

Reviewed-by: Marc Alff <marc.alff@oracle.com>