Renamed all timezone.{test,result,opt} to time_zone.
It is problematic when a feature is spelled two different ways
in filenames, for instance when searching for all tests of the
feature. The time zone tests are already in different suites, so
putting them in the same suite is not an option either.

The SQL standard spells it "time zone" (with space,) so this
renaming reflects that.

Change-Id: I72e002469e03b75b4eb4d69706edccdfacb489b1