Bug#29394407 PARALLEL DISK DATA TABLE AND TABLESPACE DDL QUERIES CAUSE MDL DEADLOCK
Problem:
--------

Parallel disk data table and tablespace queries in the *same* MySQL server
caused an MDL deadlock. Every DDL requires that an EXCLUSIVE (X) lock be
taken on the object being modified and every X lock in turn requires that
the Global Schema Lock (GSL) be acquired in NDB.

A disk data table DDL will try and acquire the following -
 - Global read lock of type INTENTION EXCLUSIVE (IX)
 - IX lock on the schema
 - Shared lock on the table
 - Backup lock of type IX
 - IX lock on the tablespace
 - Upgrade the previously acquired shared lock on the table to an EXCLUSIVE
   (X) lock
 - The X lock is granted only after the GSL has been acquired

A tablespace DDL will try and acquire the following -
 - Global read lock of type IX
 - X lock on the 'ts1' tablespace
 - The X lock is granted only after the GSL has been acquired
 - Backup lock of type IX

Assume that the table DDL has acquired an IX lock on the tablespace and is
waiting for the GSL in order to acquire an X lock on the table. At the same
time the tablespace DDL has acquired the GSL and is waiting to acquire an X
lock on the tablespace - Deadlock!

Fix:
----

Such a deadlock is detected by tracking when the GSL is acquired (and
released) during an attempt to obtain an X lock on a tablespace. When this
condition holds true (along with the other deadlock conditions), a deadlock
is assumed

Change-Id: I156f594b8746eaf9ea237e6aab41885ac6277692