Bug#29397385 WRONG RESULT FOR JSON_OBJECTAGG WINDOW FUNCTION IF DUPLICATE KEYS & UNORDERED
When the column used as a JSON key is not ordered in the window,
Item_sum_json_object::m_key_map is used to keep track of how many duplicates
we have in the frame. When moving the window frame, and the count of a key
gets down to zero, the key/value pair is removed from the JSON
object. However, the code neglects to erase the map entry for this key (now
down to one), leading to the next rows seeing a "ghost" of rows no longer in
the window frame.

Fix: remove the map entry when we no longer have occurences of that key in frame.

Change-Id: I9ee35d4edeb75afcad6e32111522c31b33fa3370