BUG#28997370 - MYSQLX PLUGIN PERFORMANCE DEGRADATION CAUSE BY LATE SENDING OF STMT_EXECUTE_OK
Post-push fix - admin_xkill toggles on machines with high load.
The patch changed behaviour of X Plugin where the state of the session
is not "stable" after sending StmtExecuteOk. Executing admin-command
kill on such client may result in two different responses.
The patch synchronizes the state of the session and the test.

RB: 21610
Reviewed by: Grzegorz Szwarc <grzegorz.szwarc@oracle.com>
Reviewed by: Tomasz Stepniak <tomasz.s.stepniak@oracle.com>