Bug#29402943 SIG11 AT SUBSELECT_SINGLE_SELECT_ENGINE::UPPER_SELECT_CONST_TABLES()
Since ORDER BY expressions may already have been bound (with rollup), when we enter
setup_order, the code didn't anticipate that the matching for the order by expression
would succeed with an Item_ref - it only anticipated Item_field.
This lead to the erroneous removal of the subquery.

Solution: add a call to real_item to the ORDER BY expression before comparing with the resolved
SELECT list item, so we get the match intended.

Change-Id: I812068986d12ade22ed5dfe93a62c156f39a1314