Bug#29328529 AGGREGATE FUNCTION RESULT IS DEPENDENT BY WINDOW IS DEFINED DIRECTLY OR AS NAMED
The result for a window without frame specification, but which inherit
a window with an ORDER BY, was wrong:

  OVER (w)   ...      w AS (ORDER BY ..)

For the first window, we a priori set the frame to RANGE and the border types (from, to)
to (UNBOUNDED_PRECEDING, UNBOUNDED_FOLLOWING). But since we inherit an ORDER BY, the
"to" border type needs adjustment. This was missing and led to us to presume we had
a window with m_static_aggregates == true.

Note that we cannot inherit an explicit frame specification from w,
but an implicit one from ORDER BY is allowed.

Note that title of this issue is somewhat misleading: a named window
reference doesn't have the parentheses:

  OVER w   ...      w AS (ORDER BY ..)

This form didn't have any issue. Adding the parenthesis defines a new window.

Change-Id: Ib157179992da48e994206c1a963f230d0231b3f2
Approved-By: GUILHEM.BICHOT <guilhem.bichot@oracle.com>