Bug#28955216: Assertion 'keyparts > 0' failed
This problem needs a number of factors to occur:
- A subquery that is converted to a semi-join.
- The subquery contains a merged derived table or a CTE.
- The derived table/CTE has a WHERE clause with an outer reference.
- The outer reference must be against an indexed column.
- The semi-join is evaluated with Materialization strategy.

(Potentially, this problem might also occur for subqueries containing
nested join clauses, but since MySQL does not support outer references
in such clauses, this is not a practical problem. The fix will,
however, also handle such cases, if outer references in join clauses
become supported).

The problem is with decorrelation of the WHERE clause of the derived
table/CTE. The predicate in the WHERE clause is not considered for
decorrelation, since the code only considers the WHERE clause of
the subquery containing the derived table/CTE. Thus, its components are
not added to the sj_inner/outer_exprs lists.

Join simplification (performed in simplify_joins()) will later move
all such join conditions to the WHERE clause. These join conditions are
later used to generate Keyuse objects that are used to generate
ref access paths during join optimization. When semi-join materialization
strategy is chosen, a new table is added to the join plan (to hold
the materialized subquery), and we rely upon sj_inner_exprs
to substitute key access from the subquery table to the materialized
table. However, since the sj_inner_exprs list has not been populated
for all fields, we may fail to do this substitution. The consequence of
this is that the old key field is kept in the join plan, which means
that it has invalid used_tables information, and thus
create_ref_for_key() will not find the key that it expects.

The best solution to the problem seems to be to recursively enter
all join nests of the subquery that may be regarded as inner joins
(in this context, a merged subquery looks like an inner join),
and decorrelate all found join conditions, in addition to the
WHERE clause, thus populating the sj_inner/outer_exprs lists so
that they are aligned with the Keyuse objects.

The fix also adds more optimizer tracing for semi-join transformation.

Reviewed by: Guilhem Bichot <guilhem.bichot@oracle.com>