Bug#28727072: BINLOG_PARSER-T CANNOT BE FOUND
Problem: binlog_parser-t needed to link with the entire server and
plugin. Consequently it had been given the EXCLUDE_FROM_ALL tag. Since
it was not built by default breakage due to changes elsewhere had gone
unnoticed.

Solution: Modify the test so that it compiles and runs without mysqld
server or plugin. Changes include:

(1) Change the include file for MySQL data types to field_types.h

(2) Copy in the definition of chars_2_decimal which is otherwise found
in the gunit_large library.

(3) Change the use of str_to_datetime() to the version which is
defined in libmytime.

(4) Create ChangePropagationTable instances in some places as they are
now needed to create Table objects.

(5) Link the test only with gunit_small, gtest, strings and mytime

(6) Make the less-than operator for rpd::Row a non-member and place it
in a separate header file so that it can also be used in unit tests.

(7) Modify test cases to also test greater than. Change all EXPECT_
macro usage so that the expected value is the first argument (to make
output more readable in case of failures), and add a new test case for
charset utf8mb4_0900_ai_ci.

Change-Id: I0a9b8a96696ec6bf7564895d04e2c3f389cb6d4a