WL#13063 ENSURES NDBMTD EXISTS WHEN REQUESTED TO USE
Modifies atrt logic to set ndbmtd as a required binary and validate
it exists when --mt arguemnt is non-zero.

atrt controls data node binary to be used via --mt argument. When 0,
do not use ndbmtd. If different, ndbmtd will be used in round-robin
(--mt=1) together with ndbd, or solo (--mt=2).
Existent implementation checked if ndbmtd existed in the path and if
not, automatically fall back to ndbd ignoring the --mt setting which was
a bug.

This change interprets the value specified in --mt argument and
accordingly sets ndbmtd as a required binary.
If --mt is non-zero, then code that checks for all required binaries
will trigger a failure if ndbmtd does not exists.

Change-Id: I688be43b951f9eb6c84b7265750e462083b7d6fb