Bug#29390106 MTR: AVOID ADDING SECONDARY ENGINE'S TEST SUITE WHEN IT ISN'T BUILT
Issue
-----
A secondary engine's test suite run will be attempted by default
and will be skipped later if the secondary engine is not built.
This process may take up some time in the test run.

Fix
---
Append such a test suite to the list of default suites run by MTR
only when the secondary engine is built.

Change-Id: Ib33021d5195c8b7f383f235dcdbf92ba025f7fbe