Bug#29441756: MTR SKIPS TESTS WHICH REQUIRE DEBUG BINARIES ON DEBUG               BUILDS WITH EXTERN
Issue:
------
When a test sources 'have_debug.inc', it requires a debug build for
it to run on. If the test is run on a non-debug build, it is skipped.
However, if the test is run using --extern on a server started separately,

Fix:
----
MTR collects the mysqld features before running any test, to check
for any support that the test may require, such as having debug builds.
This is done by running --help --verbose on the mysqld variable and
storing all the variables and their values in a hash. In this case,
the variable 'debug' is present only if it is a debug build, and
its value is '(No default value)'. With --extern, MTR executes
"SHOW VARIABLES" on the running server, and gets the values of the
variables. In this case, the value of debug is undefined. Therefore,
the check for debug support needs to check if the hash entry for
'debug' exists, and not just if it has a value.

Change-Id: I1d148b13a79a8130ca8d719408c8eca574d5ecb1