Bug#29435461 OPTIMIZE HA_NDBCLUSTER::UNPACK_RECORD()
ha_ndbcluster::unpack_record() will loop over all fields in the retrived
rows and copy them out one-by-one to the destination expected by the server.
This is expected to be a hot spot in our code and should be ensured to be as
efficient as possible.

The current implementation took a rather tedious path by using
Field::move_field_offset() back and forth to set up the source
and destination field to copy to/from. Internally that also involved
several NULL value checks which are entirely obsolete at this point
as we already knew the value to be non-NULL.

This patch optimze the common case of the field being neiter a BLOB
nor a BIT field. By using the optional 'row_offset' argument to
the Field methods being used, we could entirely avoid using ::move_field_offset().

The check for is_null was also combined into a common check for
all non-BLOB fields. The likely case of not a BLOB/BIT Field is
now implemented in the first Field-unpack code block. This also
fixed an issue with the incorrect 'is_null' method being used to
check for NULL-values. The Field::is_null() method which was used,
also checked for the TABLE::null_row flag being set, which is not
a required check at this point, as we know that a row has been found.
Replaced this with ::is_real_null() which only check the Field value
itself.

Note that in combination with bug#29435133:

JOIN_CACHE::JOIN_RECORD() MAY LEAVE A TABLE IN HAS_NULL_ROW-STATE

.. the above could cause field values to be returned as NULL instead of
the value retrieved from the storage engine.

The logic for the 'unlikely' cases of the Field being a BLOB or BIT field
was not modified.

Reviewed by: Magnus Blaudd <magnus.blaudd@oracle.com>

(cherry picked from commit a9d6e8f1a3313ebaac7de4ce2ce5215408c14a4e)