WL#13005 Implement JSON_SCHEMA_VALIDATION_REPORT
This patch implements the function JSON_SCHEMA_VALIDATION_REPORT.
The function validates a JSON document against a JSON Schema, much
like JSON_SCHEMA_VALID. The difference is that JSON_SCHEMA_VALIDATION_REPORT
returns a JSON object giving a more detailed report of the validation.
This is very useful for identifying _why_ a JSON document failed the
JSON Schema validation.

Most of the infrastructure written for JSON_SCHEMA_VALID is re-used, so
the two functions follow a lot of the same code path.

Change-Id: Ia92206c7cbd2097a2db80c12bf0c513e1431772b