WL#10162 Handle failure to send result
Schema distribution participant fails to send the result due to
error from NDB. For example running out of datamemory cause the
insert into ndb_schema_result to fail.

Problem is provoked by ndb_addnode_restart test where a table is
filled, nodes are added and then DDL is run.

Fix by falling back to "old protocol" when error to insert occurs,
this will wakeup the coordinator and works as long as participant
nodeid does not exceed number of bits in slock column. Only
drawback is that the result and message can't be transported
that way so they are lost. This problem should however be
rare and using this strategy the schema distribution protol will
continue.

Warning is written to log file when error occurs and need to be
suppressed in the ndb_addnode_restart test.

Change-Id: I3b5ca2f14ebd5c3fc4ededff48ba5aa4db67ea57