BUG#29423358: WITH MINIMAL UPGRADE GR PLUGIN BLOCKS CONNECTION TO THE SERVER
During a server upgrade with --upgrade=MINIMAL, since it only does
update the data dictionary, all other tables are not upgraded. This
makes impossible to start Group Replication because the replication
internals which depend on system tables cannot be loaded.
This new error scenario brought to light a error mishandling on the
delayed_initialization_thread, which would make the server wait
indefinitely for the Group Replication to be started when that was
impossible.

To fix the above issue, when the delayed_initialization_thread
errors out it unblock possible waiters on it.

ReviewBoard: 21688