Bug#28395115 Permission denied if grants are given through role
Root cause:
-----------
At the time of insert operation we check the permission on all the columns
through check_grant_all_columns() method.  In this method we were not checking
the global privileges granted to the user through roles. In case the user is
granted the global privileges through role then that information is stored in
the cached GRANT_INFO in field structure. We were overwriting grant->privilege
with the table access retrieved through security_context instead of first
checking the cached value in it. The cached value might be the global privilege
for the user which is set in the check_access() or table grants . It is
retrieved through the security_context.

Fix:
----
 - Moved the code up the retrieve the GRANT_INFO from the field and check for
   the cached privilege
 - Restructured the code to eliminate redundant checks.

Note - We don’t need to do the version check in  case roles because in the
       reset_for_next_command() checkout_access_map  is called in the wake of
       GRANT/REVOKE/FLISH PRIVILEGES might have been called.

Testing:
--------
 - Added the scenarios in the roles_bugs.test file
 - Added  roles_bugs_debug.test file
 - 8.0 Itch run
   --------
   push id: 14436817
   Time : 2019-03-01 14:09:15

Review:
-------
RB#21632