Bug#29394833: ASSERTION FAILED: IS_PREPARED() && !IS_OPTIMIZED()
Problem:
A condition inside an IF function which has a literal constant evaluating
to TRUE always, prompts resolver to remove the entire condition and replace
it with a TRUE item. While doing that, it checks if the subquery present
in the condition is in SELECT_LIST. If so, it cannot remove the subquery
because it could be referenced elsewhere.
For the query in question, the IF function is present in SELECT_LIST. This
prompts the resolver to not remove the subquery. It just removes the entry
in the IF function's item list. When the statement is executed, the
subquery is not resolved as part of the IF function. So asserts later
as it is not resolved.

Solution:
Check if an item is in SELECT_LIST only when it is not being resolved
from SELECT_LIST. This removes the subquery for the query along with its
entry from the IF function.

Reviewed by: Roy Lyseng <roy.lyseng@oracle.com>