Bug #29451204: CRASH IN TABLE_FUNCTION::EMPTY_TABLE()
Move creation of temporary tables into MaterializeIterator and
TableMaterializeIterator. This fixes a debug assertion failure where a
MaterializeIterator with a (materialized) table function inside it would not
create the temporary table for the table function correctly.

There's no need to have every query have to execute the tests for whether
we need materialized tables to be created or not, and moving it to the
iterators makes the overall code easier to understand.  For the time being, the
code in JOIN::prepare_result() still creates temporary tables since the old
executor needs them, but after WL #13000, JOIN::prepare_result() will no longer
be called for iterator executor queries.

This is a backport from WL #13000; the patch there was intended as a cleanup,
but fixes this bug, so it was pulled out.

Change-Id: I7c21334f889705d7c39437ad9f97a64e42e3dad3