Bug #29316814: OUT OF MEMORY WITH HIGH COLUMN COUNT RB#21710
The memory root assigned for storing the metadata was limitless.
Hence a very very large column count could deplete the available memory
and cause the computer running the client to become unstable.
Added a limit to the metadata memroot equal to the maximum network
packet size for the connection.
Now the client simply refuses to allocate more than the max_packet_size
bytes in the metadata memory root and returns an out-of-memory error
instead.
Note that the OOM error is a fatal error and the only proper action
after receiving it is to disconnect the session.
Or else you risk undefined behavior.