Bug#29479473 LOGIC FOR SHUTDOWN SERVER TIMEOUT IN MTR SCRIPTS IS FLAWED
In include/restart_mysqld.inc, we multiply $shutdown_server_timeout with a
factor (currently 6) if we are running under Valgrind or on Sparc.

However, if the test using restart_mysqld.inc does several restarts,
$shutdown_server_timeout gets multiplied by the same factor again.

Eventually, the value can overflow.

This problem also exists in rpl_stop_server.inc and shutdown_mysqld.inc

Fix the logic so that $shutdown_server_timeout remains constant.

Also, avoid a roundtrip to the server by using "--expr" instead of a select
statement to do the computation.

Change-Id: I5aaa553fd5cd31969f434b2ac140df81b83434ff