Bug#29491749: DATA DICTIONARY UPGRADE FAILING WITH ENCRYPTED MYSQL TABLESPACE
When there is a DD upgrade and the 'mysql' tablespace is encrypted, creating
the target DD tables will fail.

This patch modifies the DD table DDL statements to contain the appropriate
encryption clause according to the state of the mysql tablespace.

During the creation of the scaffolding, the encryption is irrelevant because
the preliminary hardcoded 'mysql' tablespace metadata received from the SE
does not specify encryption. However, when the scaffolding has been created,
and the persisted meta data is read back from the DD tables, we get hold of
the actual 'mysql' tablespace metadata, including the encryption attribute.
At that point, the encryption state is saved in the object table definitions,
and used to add the appropriate clause upon subsequent DDL statement execution.

Change-Id: Id04ecb5228b9964b15a51ad2d6eca3008700fad2