WL#13057: Removal of the deprecated namespace of Mysqlx.Sql.StmtExecute message
Xplugin currently supports three different values for the namespace
parameter of the StmtExecute X protocol request: 'sql', 'xplugin' and
'mysqlx'. The goal of the last one, introduced by WL#9281, was to give
the user a more self-descriptive way to encode arguments of admin
commands by using key/value pairs. To maintain backwards compatibility
the 'xplugin' namespace was kept as is but has been made deprecated.
A suitable notice message informs the user about deprecation of this
namespace.

The goal of this worklog is removing the 'xplugin' namespace from Mysqlx.Sql.StmtExecute message after 2 years of deprecations.

RB: 22555
Approved-by: Shiva Shankar <shiva.y.shankar@oracle.com>

Change-Id: Ia011a37fd115c8db2c81ecdc082e30e012e484e2