Bug#29474188 MISSING RESULT ROWS FROM A SORTED RESULT SPECIFYING A 'LIMIT'
The NdbCluster handler interface did not implement the handler method
::index_next_same(), thus the default implementation in
handler::index_next_same() was used.

That implementation assumed that the handler was not able to stop
scanning when at the end of the specified range, thus it called
key_cmp_if_same() to detect when the 'end' had been found.

That depended on that all Fields being part of the 'key'
being compared had been set in the read_set, and returned by the
handler. However, there are no such guarantee, and neither do we want
to return extra fields not requested by the query.

As the end-key has already been sent to NdbCluster when we opened
the index scan, the entire key_cmp_if_same() is obsolete at the client side.

This patch override the index_next_same() implementation for
ha_ndbcluster. All it actually need to do is to request
the 'next_result()' from the sorted index cursor.

Note that also the InnoDb and MyISAM handlers override index_next_same().

Reviewed by: Magnus Blaudd <magnus.blaudd@oracle.com>