Bug #29154379: CRASHING IN SRV_MBR_PRINT
Fix an issue where multi-table UPDATE would not properly reset the NULL
row status for tables it was trying to update. This would lead to generated
columns getting the wrong value when looking for the row to update in
secondary indexes.

This wasn't a problem since the old executor, since it would reset the NULL
row status after processing each and every NULL-complemented row. However,
the iterator executor is slightly lazier and only changes said status after a
new outer row comes in, which means it could linger on if the very last row
in the result set was NULL-complemented.

Change-Id: Ifab3c1602794da1710bbd21abce0e39228c4fc66