Bug#29512411 COMPILE FATAL CON EXTENDED WERROR FORMAT SPECIFIER
The sourced memcached source files generates several warnings when
compiled with modern compilers. In order to avoid that this causes
compilation to fail when compiling with the compiler set to treat
warnings as fatal error, the compiler is instructed to ignore warnings
for these files. The technique used to instruct the compiler is
unfortunately not entirely correct and causes faulty instructions to be
sent to the compiler.

Replace the replace of values in CMAKE_C_FLAGS with append of the
instructions to turn off treating warnings as fatal error. Also remove
turning on more warnings just after warnings have been silenced.

This patch does not supress any warnings which was not previously
supressed.

Change-Id: I16601844f5106a7e888af4ca5f4e9d82895e98fe