Bug#29020745: INNODB TABLESPACES WITH SUBPARTITIONS DO NOT CONTAIN TABLE SDI OBJECT
Problem: When using subpartitioning the table SDI would not be stored
in any of the .ibd files. Root cause was that the code for writing
sdis were not updated when the subpartitions were placed in a separate
container in the dd::Table object, and the leaf_partitions vector
introduced.

Solution: Change the sdi writing code so that it looks for tablespace
ids in partitions and subpartitions as well as partition indexes, and
uses the first tablespace id which is different from INVALID_OBJECT_ID.

It was necessary update an assert in dict_sdi_set() to use
leaf_partitions when checking for valid se_private_id members.

Additionally, address some technological debt which had accumulated:

(1) Factor out the code used to find tablespace ids in a dd::Table
graph and use it both for sdi writing and encryption checking.

(2) Simplify and move the definition of sdi_key_t and sdi_vector_t
from dd/types/tablespace.h to handler.h and out of the dd namespace.
Move two constants from dd/types/tablespace.h to include/dict0sdi.h
since they were not part of the handlertion API. Since the types moved
out of dd:: are part of the handlerton api, this changes also
affects NDB, and memcached.

Change-Id: I4bd547a179aa5d11e13e79d21b01473b66612e85