Bug#29487181 INCORRECT 'ROWS' AND 'FILTERED' ESTIMATE IN 'ORDER BY ... LIMIT .' QUERIES
When optimizing a joined table query, where also an 'order by ... limit ..' clause was
specified, the optimizer modified the 'select_limit' on the first table by taking the
'fanout' of the entire join into consideration.

This 'fanout' was calculated using the 'filter_effect' on each table. 'filter_effect' could
be set to 'COND_FILTER_STALE' (=-1) to indicate an 'unknown' value.

This was not checked for by the optimizer, resulting in a negative 'fanout', and thus
the unexpected: 'rows = 0'.

This patch checks for a table having an unknown 'filter_effect'. If found, the
'fanout' is also assumed to be 'unknown', and the 'rows' to be selected from the
first table is then not modified by the fanout of the join query.

Reviewed by: Knut Anders Hatlen <knut.hatlen@oracle.com>