Bug#29513265: I_S VIEW META DATA NOT RECREATED WHEN UNDERLYING DD TABLES ARE CHANGED
When the DD tables are changed in a DD upgrade, the I_S views should have
their meta data recreated in case e.g. a column type, collation, size etc.
is changed.

This patch makes the server keep track of the fact that there has been a
DD upgrade, and uses this when the I_S is initialized to make sure the
views are re-created if there has been a DD upgrade.

Change-Id: I7e25a9d2ee6f2953ff2e4116bcb3d9a9c4c96bf5