Bug#28884359: LAST_QUERY_COST IS NEGATIVE AFTER ORDER BY/LIMIT QUERY
The code that adjusts the query cost after sort elimination could
compensate for sorts that had not been considered in the original cost
estimate, and therefore cause too low estimates, sometimes even
negative ones.

The sort cost is added to the total cost in
Optimize_table_order::consider_plan(), and subtracted from the total
cost in JOIN::optimize() if it is found that sorting is not needed.

If consider_plan() finds that sorting is not needed, it does not add
the sort cost to the total cost. However, it still records a sort cost
in JOIN::sort_cost, and this cost is set to the sort cost of the
previously considered candidate plan. Later, JOIN::optimize() sees
that the JOIN object has an associated sort cost, and subtracts that
cost from the total cost, since it also sees that no sort is needed.
Since the sort cost comes from a candidate plan that is not the same
as the chosen plan, the estimate is wrong.

The fix is to make consider_plan() set JOIN::sort_cost to 0 if no sort
cost is added to the total cost, so that the mistaken adjustment of
the cost in JOIN::optimize() is not performed.

Change-Id: I734653837ab1dd543f0e59e31bc20146091471fe