Bug #28341329: COMPONENT OPTIONS WITH --LOOSE PREFIX AREN'T CONSIDERED  RB#21745
The system variable service register method was trying to initialize the
values for the newly registered variable from the command line, but was
using the wrong argc/argv. Instead of using the original one supplied to
the service (as e.g. the plugin variables do) it was using whatever was
"remaining" from the original parsing.

Fixed by using the same argv/argc as for the plugin variables.
But since components use a singe function this needed to use a different
source based on the state the server is in.
Test case added.