Bug#29493830 CONST'IFIED OUTER JOIN RETURN INCORRECT RESULTS
When a table is const optimized, any predicate terms on the joins ON-condition
is also evaluated on the const'ified tables to make sure that the row
qualifies on the ON-condition.

The ON-conditions might have been transformed to a 'multiple equality Item',
represented as a Item_equal object at this point of the optimization process.
Effectively the ON-condition is evaluated in Item_equal::val_int(), which
will only check the equality for any Item_field where the fields table
has been marked as a 'const_table' - comparison of other field values
are intentionaly skipped.

It turned out that the optimizer failed to mark a table as a 'const_table'
when it was handled as a 'impossible ON condition'. In this case the resulting
row from this table will always be a NULL-extended row.

This patch set the 'const_table' flag together with setting the
row as a NULL-row.

Patch also change some explain output from other test cases.
These changes are exclusively related to table being
'Impossible ON-condition' optimized, where the resulting NULL values from
this row is now const propagated into the rows Field values.
(Just like any other const-row)

Reviewed by: Roy Lyseng <roy.lyseng@oracle.com>