Bug#29512245 PERFORMANCE SCHEMA SERVICES USING HOSTNAME OF LENGTH 60
Context
-------

The performance schema exposes several services, including:
- pfs_notification
- pfs_resource_group
- psi_thread_v1
- psi_thread_v2

All of these services exposes APIs related to thread attributes,
as in struct PSI_thread_attrs_v1.

This structure contains the following member:

define PSI_HOSTNAME_LENGTH (60)
char m_hostname[PSI_HOSTNAME_LENGTH];

Problem
-------

This limitation of 60 characters for a host name is historical in MySQL,
due to how user@host accounts are named, and needs to be expanded to 255
characters, to support any DNS host name.

Simply changing the existing member to:
  char PSI_thread_attrs_v1::m_hostname[255]
is an api change that breaks binary compatibility of an exposed service,
and can not be done.

Solution
--------

Create new versions for all the services impacted,
all named "version 3" for consistency, to perform the change.

In this version, the thread attributes structure is modified as follows:
  char PSI_thread_attrs_v3::m_hostname[255]

The new services defined are:
- pfs_notification_v3
- pfs_resource_group_v3
- psi_thread_v3

Note that the following services are discontinued:
- pfs_notification
- pfs_resource_group
- psi_thread_v1
- psi_thread_v2

The rational for discontinuing old services is as follows:

A)

Preserving support for it may appear to work for a while,
but these services are flawed because they will all truncate a host name
to 60 characters.

This affects in particular environment using containers, like docker,
which typically uses very long host names.

B)

This truncation is, by definition of the old services APIs,
impossible to fix in the old services.

The proper (and only) fix is to use the new version provided instead,
and this does require to recompile the client component
code to call the _v3 services.

C)

Assume it is feasible to keep multiple versions of the notification services,
like:
- pfs_notification
- pfs_notification_v3

Now, when threads events needs to be notified,
every notification service must be called, in every versions.

This causes un necessary performance overhead.

Approved by: Chris Powers <chris.powers@oracle.com>