Bug#29491083: CRASH IN ITEM_FUNC_LIKE::CHECK_COVERING_PREFIX_KEYS()
Item::is_null() returned an incorrect result for some expressions
that were comparing rows, and this could in some cases cause the
server to exit.

The problem is that Item_bool_func2::is_null() does not evaluate the
expression that it represents. It only checks if any of its arguments
evaluates to NULL. This is not correct for comparison operators that
take rows as operands, as one cannot determine if they evaluate to
NULL just by checking if their operands contain NULL values.

The fix is to make Item_func_comparison::is_null() use the fast path
only if its operands are scalar. If the operands are rows, the
operator is fully evaluated in order to see if it evaluates to NULL.

Additionally, change a check for Item::is_null() to Item::null_value
in Item_func_like::check_covering_prefix_keys(). The Item has already
been evaluated, so there is no need to evaluate it again to find out
if it is NULL.

Change-Id: I87cdc1c15f4a1cc56c21eae47ed6ecac1f2a8732